Save the Queen!
This project contains two python files, the main file and its corresponding unit test. In queens.py, the program evaluates the safety of the queen, similar to chess, and will warn the queen based on if there is another queen directly next to them (either directly above, below, to the left, right, or diagonally). Included is also a QueenSafety.png which depicts threatening positions. There are functions that allow a user to add queens to the board, which is saved in the QueensState class __init__ method, and you are also able to remove a queen. You can then check if there is a queen in a certain position, if there are any unsafe queens.